refactor(#41): 예외 형식 통일화
