+++ src/main/java/cinebox/controller/ScreenController.java
@@ -1,13 +1,20 @@
 package cinebox.controller;
 
-import cinebox.dto.ScreenRequest;
-import cinebox.dto.ScreenResponseDto;
-import cinebox.service.ScreenService;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import cinebox.dto.request.ScreenRequest;
+import cinebox.dto.response.ScreenResponse;
+import cinebox.service.ScreenService;
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequestMapping("/api/screen")
@@ -18,18 +25,18 @@ public class ScreenController {
     // 상영 정보 추가
     @PostMapping
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public ResponseEntity<ScreenResponseDto> createScreen(@RequestBody ScreenRequest request) {
-        ScreenResponseDto responseDto = screenService.createScreen(request);
+    public ResponseEntity<ScreenResponse> createScreen(@RequestBody ScreenRequest request) {
+        ScreenResponse responseDto = screenService.createScreen(request);
         return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
     }
 
     // 상영 정보 수정
     @PutMapping("/{screenId}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public ResponseEntity<ScreenResponseDto> updateScreen(
+    public ResponseEntity<ScreenResponse> updateScreen(
             @PathVariable("screenId") Long screenId,
             @RequestBody ScreenRequest request) {
-        ScreenResponseDto updatedScreen = screenService.updateScreen(screenId, request);
+        ScreenResponse updatedScreen = screenService.updateScreen(screenId, request);
         return ResponseEntity.ok(updatedScreen);
     }
